It's getting too late, and I need to go to bed.
An idea here is to write an assembler before writing a compiler.  Not only will this make writing the compiler easier, but it should also provide a way to test the VM well before the compiler is ever to a point where it can be tested.  Furthermore, I can use the assembler to figure out the best way to the handle the calling convention for the compiler.

This is not an assembler in traditional sense.  Rather, it's code that can take a simple data-structure (a sequence of instruction instance) and produce from that an executable program buffer.

Ultimately, blocks of individual instruction sequences will get created independently, then strung altogether in the right order for final assembly.  This, in some sense, could be thought of as a flattening of the syntax tree, or perhaps some intermediate tree-like data-structure that is derived from the syntax tree.

The syntax tree should get grammar checked before moving on to the next stage.  Errors can occur at any of the stages, such as tokenizing, parsing, lexical analysis, etc.