Write code to test new code re-usability feature, and then realize, hilariously, that it will absolutely never work.
A function floating around in the VM has to be more than just an address, or the address value has to hold open a reference to the code buffer, which needs to become a GCCollectable.  Furthermore, a jump address has to be more than just an offset, but also indicate the code buffer, so that you can do intra-code-buffer jumps, not just inter-code-buffer jumps.  Of course, all of that is just one idea.  Maybe there's a cleaner way?