Instruction printing is no longer a debug method, because it's used in a release build for showing the disassembly.
Also, realize a bug in the interpreter...

Powder: a = func(x, y) { return x + y; }
Powder: output(a(1, 2));

This will crash.  What's the best way to handle this?  Does the function value need to maintain a reference to the executable?  Even so, how does the jump instruction jump to it?